Moved refresh token into its own class.
